+++ src/app/data/api.service.ts
@@ -13,13 +13,15 @@ import {
 } from './models';
 
 import { SettingsService } from './settings.service';
-import { asNumber, data_dust_events, daysUntil, diffNumbers, nowAtEvent, static_dust_events } from '../utils/utils';
+import { asNumber, data_dust_events, daysUntil, diffNumbers, isAfter, nowAtEvent, static_dust_events } from '../utils/utils';
 import { DbService } from './db.service';
 import { Capacitor } from '@capacitor/core';
 import { App } from '@capacitor/app';
 import { environment } from 'src/environments/environment';
 import { getCachedImage } from './cache-store';
 import { distance } from '../map/map.utils';
+
+export type DownloadResult = 'success' | 'error' | 'already-updated';
 
 interface Version {
   version: string;
@@ -155,10 +157,28 @@ export class ApiService {
         dataset.subTitle = `${dataset.title} is happening now`;
       }
     }
-    return datasets
+    const list = datasets
       .filter((d) => d.active || inactive)
       .filter((d) => this.byType(d, filter))
-      .sort((a, b) => diffNumbers(a.dist, b.dist));
+      .sort((a, b) => this.sortEvents(a, b, filter));
+    if (list.length > 1 && filter == 'all' && isAfter(new Date(list[0].start), new Date(list[1].start))) {
+      // Make Burning Man first if the closest regional is after it.
+      const tmp = { ...list[0] };
+      list[0] = list[1];
+      list[1] = tmp;
+    }
+    return list;
+  }
+
+  private sortEvents(a: Dataset, b: Dataset, filter: DatasetFilter): number {
+    const diff = diffNumbers(a.dist, b.dist);
+    if (filter === 'all') {
+      if (a.id.startsWith('ttitd')) {
+        return -1;
+      }
+    }
+
+    return diff;
   }
 
   private byType(a: Dataset, filter: DatasetFilter): boolean {
@@ -200,7 +220,7 @@ export class ApiService {
     selected: Dataset | undefined,
     force: boolean,
     downloadSignal: WritableSignal<string>,
-  ): Promise<boolean> {
+  ): Promise<DownloadResult> {
     let dataset = '';
     let nextRevision: Revision | undefined;
     let myRevision: Revision | undefined;
@@ -247,15 +267,15 @@ export class ApiService {
         console.log(
           `Will not download data for ${dataset} as it is already at revision ${nextRevision.revision} and version is ${currentVersion}`,
         );
-        return true;
+        return 'already-updated';
       }
     } catch (err) {
       console.log('Possible CORs error as document location is ', document.location.href);
       console.error('Unable to download', err);
-      return false;
+      return 'error';
     }
     console.log(`Will attempt download to ${JSON.stringify(nextRevision)}`);
-    if (!nextRevision) return false;
+    if (!nextRevision) return 'error';
 
     downloadSignal.set(selected ? selected.title : ' ');
     const currentVersion = await this.getVersion();
@@ -296,7 +316,7 @@ export class ApiService {
     if (this.badData(events, art, camps)) {
       console.error(`Download has no events, art or camps and has failed.`);
       downloadSignal.set('');
-      return false;
+      return 'error';
     } else {
       console.log(`Data passed checks for events, art and camps`);
     }
@@ -307,6 +327,6 @@ export class ApiService {
     await this.dbService.writeData(dataset, Names.map, map);
 
     downloadSignal.set('');
-    return true;
+    return 'success';
   }
 }