+++ changelog.md
@@ -1,5 +1,12 @@
 ## Changes
 
+## 2.21
+
+- Option to download update added to the more menu
+- Fix missing add to favorites for art
+- Show Burning Man first in the list if it starts earlier than the closest regional burn
+- Geolocation check improvements for skipping location and enabling location
+
 ## 2.20
 
 - Prevent downloading updates when the event has started and cell network is used